Move view logic
